% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_atk_recipe.R
\name{add_recipe}
\alias{add_recipe}
\title{Populate a directory structure for recipes}
\usage{
add_recipe(title, source)
}
\arguments{
\item{title}{recipe title, , snake case, all lowercase letters}

\item{source}{character string indicating issue or book where recipe was found. For example, "Cook's Country December/January 2021"}
}
\value{

}
\description{
Standardize the hierarchy of folders and files to allow iterative recipe assembly and indexing
}
\details{
Creates a main folder containing two blank files:
\enumerate{
\item A blank csv file in which to add ingredients
\item A blank markdown file titled `\<recipe title\>_directions.md.
A png image should be added to the folder to complete the recipe, which will be stored as a list object in the package.
}
}
