% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_shopping_list.R
\name{create_shopping_list}
\alias{create_shopping_list}
\title{Generate a shopping list}
\usage{
create_shopping_list(path, lump = TRUE)
}
\arguments{
\item{path}{path to a spreadsheet file (supported formats include .csv, .xls,
.xlsx)}

\item{lump}{whether to combine all usages of a given ingredient into one
quantity. If different units are used to measure the same ingredient, the
instances will be kept separate. For example, "1 lb tomatoes" and "6
tomatoes" will remain separate entries.}
}
\value{
A tibble arranged by department and item
}
\description{
Takes a spreadsheet file containing meals and organizes the items by store
layout. Optionally sums all uses of a given ingredient from multiple meals
into a single quantity to procure.
}
