% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_csv_list.R
\name{manual_csv_list}
\alias{manual_csv_list}
\title{Populate a blank csv file with column names}
\usage{
manual_csv_list(date = Sys.Date())
}
\arguments{
\item{date}{date of planned shopping trip}
}
\value{
file silently written to disk in the \code{inst/shopping_lists} directory
}
\description{
Useful for making manual lists (i.e. not from specific meals)
}
\details{
\lifecycle{experimental}
}
